+++ core/DataTable/Row.php
@@ -19,7 +19,6 @@ use Piwik\Metrics;
  * DataTable rows contain columns, metadata and a subtable ID. Columns and metadata
  * are stored as an array of name => value mappings.
  *
- *
  * @api
  */
 class Row implements \ArrayAccess, \IteratorAggregate
@@ -34,23 +33,17 @@ class Row implements \ArrayAccess, \IteratorAggregate
         'full_url' => true // column used w/ old Piwik versions,
     );
 
-    /**
-     * This array contains the row information:
-     * - array indexed by self::COLUMNS contains the columns, pairs of (column names, value)
-     * - (optional) array indexed by self::METADATA contains the metadata,  pairs of (metadata name, value)
-     * - (optional) integer indexed by self::DATATABLE_ASSOCIATED contains the ID of the DataTable associated to this row.
-     *   This ID can be used to read the DataTable from the DataTable_Manager.
-     *
-     * @var array
-     * @see constructor for more information
-     * @ignore
-     */
-    public $c = array();
-
-    private $subtableIdWasNegativeBeforeSerialize = false;
-
     // @see sumRow - implementation detail
     public $maxVisitsSummed = 0;
+
+    private $columns = array();
+    private $metadata = array();
+    private $isSubtableLoaded = false;
+
+    /**
+     * @internal
+     */
+    public $subtableId = null;
 
     const COLUMNS = 0;
     const METADATA = 1;
@@ -73,60 +66,33 @@ class Row implements \ArrayAccess, \IteratorAggregate
      */
     public function __construct($row = array())
     {
-        $this->c[self::COLUMNS] = array();
-        $this->c[self::METADATA] = array();
-        $this->c[self::DATATABLE_ASSOCIATED] = null;
-
         if (isset($row[self::COLUMNS])) {
-            $this->c[self::COLUMNS] = $row[self::COLUMNS];
+            $this->columns = $row[self::COLUMNS];
         }
         if (isset($row[self::METADATA])) {
-            $this->c[self::METADATA] = $row[self::METADATA];
+            $this->metadata = $row[self::METADATA];
         }
-        if (isset($row[self::DATATABLE_ASSOCIATED])
-            && $row[self::DATATABLE_ASSOCIATED] instanceof DataTable
-        ) {
-            $this->setSubtable($row[self::DATATABLE_ASSOCIATED]);
-        }
-    }
-
-    /**
-     * Because $this->c[self::DATATABLE_ASSOCIATED] is negative when the table is in memory,
-     * we must prior to serialize() call, make sure the ID is saved as positive integer
-     *
-     * Only serialize the "c" member
-     * @ignore
-     */
-    public function __sleep()
-    {
-        if (!empty($this->c[self::DATATABLE_ASSOCIATED])
-            && $this->c[self::DATATABLE_ASSOCIATED] < 0
-        ) {
-            $this->switchFlagSubtableIsLoaded();
-            $this->subtableIdWasNegativeBeforeSerialize = true;
-        }
-        return array('c');
-    }
-
-    /**
-     * Must be called after the row was serialized and __sleep was called.
-     * @ignore
-     */
-    public function cleanPostSerialize()
-    {
-        if ($this->subtableIdWasNegativeBeforeSerialize) {
-            $this->switchFlagSubtableIsLoaded();
-            $this->subtableIdWasNegativeBeforeSerialize = false;
+        if (isset($row[self::DATATABLE_ASSOCIATED])) {
+            if ($row[self::DATATABLE_ASSOCIATED] instanceof DataTable) {
+                $this->setSubtable($row[self::DATATABLE_ASSOCIATED]);
+            } else {
+                $this->subtableId = $row[self::DATATABLE_ASSOCIATED];
+            }
         }
     }
 
     /**
-     * note: also used by unit tests
+     * Used when archiving to serialize the Row's properties.
+     * @return array
      * @ignore
      */
-    public function switchFlagSubtableIsLoaded()
+    public function export()
     {
-        $this->c[self::DATATABLE_ASSOCIATED] = -1 * $this->c[self::DATATABLE_ASSOCIATED];
+        return array(
+            self::COLUMNS => $this->columns,
+            self::METADATA => $this->metadata,
+            self::DATATABLE_ASSOCIATED => $this->subtableId,
+        );
     }
 
     /**
@@ -135,9 +101,10 @@ class Row implements \ArrayAccess, \IteratorAggregate
      */
     public function __destruct()
     {
-        if ($this->isSubtableLoaded()) {
-            Manager::getInstance()->deleteTable($this->getIdSubDataTable());
-            $this->c[self::DATATABLE_ASSOCIATED] = null;
+        if ($this->isSubtableLoaded) {
+            Manager::getInstance()->deleteTable($this->subtableId);
+            $this->subtableId = null;
+            $this->isSubtableLoaded = false;
         }
     }
 
@@ -175,10 +142,11 @@ class Row implements \ArrayAccess, \IteratorAggregate
      */
     public function deleteColumn($name)
     {
-        if (!$this->hasColumn($name)) {
+        if (!array_key_exists($name, $this->columns)) {
             return false;
         }
-        unset($this->c[self::COLUMNS][$name]);
+
+        unset($this->columns[$name]);
         return true;
     }
 
@@ -190,11 +158,12 @@ class Row implements \ArrayAccess, \IteratorAggregate
      */
     public function renameColumn($oldName, $newName)
     {
-        if (isset($this->c[self::COLUMNS][$oldName])) {
-            $this->c[self::COLUMNS][$newName] = $this->c[self::COLUMNS][$oldName];
+        if (isset($this->columns[$oldName])) {
+            $this->columns[$newName] = $this->columns[$oldName];
         }
+
         // outside the if () since we want to delete nulled columns
-        unset($this->c[self::COLUMNS][$oldName]);
+        unset($this->columns[$oldName]);
     }
 
     /**
@@ -205,11 +174,11 @@ class Row implements \ArrayAccess, \IteratorAggregate
      */
     public function getColumn($name)
     {
-        if (!isset($this->c[self::COLUMNS][$name])) {
+        if (!isset($this->columns[$name])) {
             return false;
         }
 
-        return $this->c[self::COLUMNS][$name];
+        return $this->columns[$name];
     }
 
     /**
@@ -221,12 +190,12 @@ class Row implements \ArrayAccess, \IteratorAggregate
     public function getMetadata($name = null)
     {
         if (is_null($name)) {
-            return $this->c[self::METADATA];
+            return $this->metadata;
         }
-        if (!isset($this->c[self::METADATA][$name])) {
+        if (!isset($this->metadata[$name])) {
             return false;
         }
-        return $this->c[self::METADATA][$name];
+        return $this->metadata[$name];
     }
 
     /**
@@ -238,7 +207,7 @@ class Row implements \ArrayAccess, \IteratorAggregate
      */
     public function hasColumn($name)
     {
-        return array_key_exists($name, $this->c[self::COLUMNS]);
+        return array_key_exists($name, $this->columns);
     }
 
     /**
@@ -254,7 +223,7 @@ class Row implements \ArrayAccess, \IteratorAggregate
      */
     public function getColumns()
     {
-        return $this->c[self::COLUMNS];
+        return $this->columns;
     }
 
     /**
@@ -265,10 +234,7 @@ class Row implements \ArrayAccess, \IteratorAggregate
      */
     public function getIdSubDataTable()
     {
-        return !is_null($this->c[self::DATATABLE_ASSOCIATED])
-            // abs() is to ensure we return a positive int, @see isSubtableLoaded()
-            ? abs($this->c[self::DATATABLE_ASSOCIATED])
-            : null;
+        return $this->subtableId;
     }
 
     /**
@@ -278,14 +244,24 @@ class Row implements \ArrayAccess, \IteratorAggregate
      */
     public function getSubtable()
     {
-        if ($this->isSubtableLoaded()) {
+        if ($this->isSubtableLoaded) {
             try {
-                return Manager::getInstance()->getTable($this->getIdSubDataTable());
+                return Manager::getInstance()->getTable($this->subtableId);
             } catch(TableNotFoundException $e) {
                 // edge case
             }
         }
         return false;
+    }
+
+    /**
+     * @param int $subtableId
+     * @ignore
+     */
+    public function setNonLoadedSubtableId($subtableId)
+    {
+        $this->subtableId = $subtableId;
+        $this->isSubtableLoaded = false;
     }
 
     /**
@@ -298,7 +274,7 @@ class Row implements \ArrayAccess, \IteratorAggregate
      */
     public function sumSubtable(DataTable $subTable)
     {
-        if ($this->isSubtableLoaded()) {
+        if ($this->isSubtableLoaded) {
             $thisSubTable = $this->getSubtable();
         } else {
             $this->warnIfSubtableAlreadyExists();
@@ -320,9 +296,9 @@ class Row implements \ArrayAccess, \IteratorAggregate
      */
     public function setSubtable(DataTable $subTable)
     {
-        // Hacking -1 to ensure value is negative, so we know the table was loaded
-        // @see isSubtableLoaded()
-        $this->c[self::DATATABLE_ASSOCIATED] = -1 * $subTable->getId();
+        $this->subtableId = $subTable->getId();
+        $this->isSubtableLoaded = true;
+
         return $subTable;
     }
 
@@ -335,8 +311,7 @@ class Row implements \ArrayAccess, \IteratorAggregate
     {
         // self::DATATABLE_ASSOCIATED are set as negative values,
         // as a flag to signify that the subtable is loaded in memory
-        return !is_null($this->c[self::DATATABLE_ASSOCIATED])
-        && $this->c[self::DATATABLE_ASSOCIATED] < 0;
+        return $this->isSubtableLoaded;
     }
 
     /**
@@ -344,7 +319,8 @@ class Row implements \ArrayAccess, \IteratorAggregate
      */
     public function removeSubtable()
     {
-        $this->c[self::DATATABLE_ASSOCIATED] = null;
+        $this->subtableId = null;
+        $this->isSubtableLoaded = false;
     }
 
     /**
@@ -354,7 +330,7 @@ class Row implements \ArrayAccess, \IteratorAggregate
      */
     public function setColumns($columns)
     {
-        $this->c[self::COLUMNS] = $columns;
+        $this->columns = $columns;
     }
 
     /**
@@ -365,7 +341,7 @@ class Row implements \ArrayAccess, \IteratorAggregate
      */
     public function setColumn($name, $value)
     {
-        $this->c[self::COLUMNS][$name] = $value;
+        $this->columns[$name] = $value;
     }
 
     /**
@@ -376,7 +352,7 @@ class Row implements \ArrayAccess, \IteratorAggregate
      */
     public function setMetadata($name, $value)
     {
-        $this->c[self::METADATA][$name] = $value;
+        $this->metadata[$name] = $value;
     }
 
     /**
@@ -388,13 +364,13 @@ class Row implements \ArrayAccess, \IteratorAggregate
     public function deleteMetadata($name = false)
     {
         if ($name === false) {
-            $this->c[self::METADATA] = array();
+            $this->metadata = array();
             return true;
         }
-        if (!isset($this->c[self::METADATA][$name])) {
+        if (!isset($this->metadata[$name])) {
             return false;
         }
-        unset($this->c[self::METADATA][$name]);
+        unset($this->metadata[$name]);
         return true;
     }
 
@@ -407,7 +383,7 @@ class Row implements \ArrayAccess, \IteratorAggregate
      */
     public function addColumn($name, $value)
     {
-        if (isset($this->c[self::COLUMNS][$name])) {
+        if (isset($this->columns[$name])) {
             throw new Exception("Column $name already in the array!");
         }
         $this->setColumn($name, $value);
@@ -443,7 +419,7 @@ class Row implements \ArrayAccess, \IteratorAggregate
      */
     public function addMetadata($name, $value)
     {
-        if (isset($this->c[self::METADATA][$name])) {
+        if (isset($this->metadata[$name])) {
             throw new Exception("Metadata $name already in the array!");
         }
         $this->setMetadata($name, $value);
@@ -537,7 +513,7 @@ class Row implements \ArrayAccess, \IteratorAggregate
      */
     public function sumRowMetadata($rowToSum)
     {
-        if (!empty($rowToSum->c[self::METADATA])
+        if (!empty($rowToSum->metadata)
             && !$this->isSummaryRow()
         ) {
             // We shall update metadata, and keep the metadata with the _most visits or pageviews_, rather than first or last seen
@@ -545,10 +521,10 @@ class Row implements \ArrayAccess, \IteratorAggregate
                 // Old format pre-1.2, @see also method doSumVisitsMetrics()
                 $rowToSum->getColumn('nb_actions') || $rowToSum->getColumn('nb_visits'));
             if (($visits && $visits > $this->maxVisitsSummed)
-                || empty($this->c[self::METADATA])
+                || empty($this->metadata)
             ) {
                 $this->maxVisitsSummed = $visits;
-                $this->c[self::METADATA] = $rowToSum->c[self::METADATA];
+                $this->metadata = $rowToSum->metadata;
             }
         }
     }
@@ -704,12 +680,12 @@ class Row implements \ArrayAccess, \IteratorAggregate
     }
 
     public function getIterator() {
-        return new \ArrayIterator($this->c[self::COLUMNS]);
+        return new \ArrayIterator($this->columns);
     }
 
     private function warnIfSubtableAlreadyExists()
     {
-        if (!is_null($this->c[self::DATATABLE_ASSOCIATED])) {
+        if (!is_null($this->subtableId)) {
             Log::warning(
                 "Row with label '%s' (columns = %s) has already a subtable id=%s but it was not loaded - overwriting the existing sub-table.",
                 $this->getColumn('label'),